// src/app/layouts/dashboard/pages/courses/models/course.ts
export interface Course {
    id: number;
    nombre: string;
    carrera: string;
    categoria: string;
    nivel: string;
    dedicacion: string;
    precioDeLista: number;
    descuento: number;
    precioFinal: number;
    detalles: string;
    rangoDeFechas: string;
    dias: string;
    horario: string;
    profesor: string;    
}